requirements.txt
# Backend Dependencies
# Note: Using flexible versions for Python 3.13 compatibility
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
# psycopg2-binary>=2.9.9  # Optional: Only needed for PostgreSQL. Uncomment if using PostgreSQL in production.
python-dotenv>=1.0.0
httpx>=0.25.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.9.0
pydantic-settings>=2.1.0
alembic>=1.13.1

# Vector DB (Day 6-7)
chromadb>=1.4.0

# AI/LLM (Day 6-7)
openai>=1.12.0

# Agents (for Week 2 - optional, install when needed)
# crewai>=0.1.32

# Utilities
python-multipart==0.0.6
email-validator==2.1.0

